Create ArxivPaperCluster
The simple idea is to find how to cluster papers posted in arXiv. In order to do this, the pdfs are downloaded then converted into text, after which the text is  processed to find the word counts, while only keeping words of more than three letters which appear in a dictionary. This data can then be used to fit a basic clustering algorithm.